+++ src/rpc/mining.cpp
@@ -34,6 +34,7 @@
 #include <rpc/util.h>
 #include <script/descriptor.h>
 #include <script/script.h>
+#include <script/standard.h>
 #include <shutdown.h>
 #include <timedata.h>
 #include <txmempool.h>
@@ -747,7 +748,8 @@ static RPCHelpMan getblocktemplate() {
                      {
                          {RPCResult::Type::OBJ,
                           "minerfund",
-                          "information related to the coinbase miner fund",
+                          "information related to the coinbase miner fund."
+                          "This will NOT be set if -simplegbt is enabled",
                           {
 
                               {RPCResult::Type::ARR,
@@ -766,10 +768,10 @@ static RPCHelpMan getblocktemplate() {
                          {RPCResult::Type::OBJ,
                           "stakingrewards",
                           "information related to the coinbase staking reward "
-                          "output, only set after the Nov. 15, 2023 upgrade "
-                          "activated and the -avalanchestakingrewards option "
-                          "is "
-                          "enabled",
+                          "output, only set if the -avalanchestakingrewards "
+                          "option is enabled and if the node is able to "
+                          "determine a winner. This will NOT be set if "
+                          "-simplegbt is enabled",
                           {
                               {RPCResult::Type::OBJ,
                                "payoutscript",
@@ -847,6 +849,34 @@ static RPCHelpMan getblocktemplate() {
                          {RPCResult::Type::STR_HEX, "nexttarget",
                           "The real-time target in compact format"},
                      }},
+                    {RPCResult::Type::OBJ,
+                     "minerfund",
+                     "information related to the coinbase miner fund."
+                     "This will ONLY be set if -simplegbt is enabled",
+                     {
+                         {RPCResult::Type::STR_HEX, "script",
+                          "The scriptpubkey for the miner fund output in "
+                          "hex format"},
+                         {RPCResult::Type::STR_AMOUNT, "amount",
+                          "The minimum value the miner fund output must "
+                          "pay in satoshis"},
+
+                     }},
+                    {RPCResult::Type::OBJ,
+                     "stakingrewards",
+                     "information related to the coinbase staking reward "
+                     "output, only set if the -avalanchestakingrewards "
+                     "option is enabled and if the node is able to "
+                     "determine a winner. This will ONLY be set if "
+                     "-simplegbt is enabled",
+                     {
+                         {RPCResult::Type::STR_HEX, "script",
+                          "The scriptpubkey for the staking reward "
+                          "output in hex format"},
+                         {RPCResult::Type::STR_AMOUNT, "amount",
+                          "The minimum value the staking reward output must "
+                          "pay in satoshis"},
+                     }},
                 }},
         },
         RPCExamples{HelpExampleCli("getblocktemplate", "") +
@@ -855,6 +885,7 @@ static RPCHelpMan getblocktemplate() {
             const JSONRPCRequest &request) -> UniValue {
             NodeContext &node = EnsureAnyNodeContext(request.context);
             ChainstateManager &chainman = EnsureChainman(node);
+            ArgsManager &argsman = EnsureArgsman(node);
             LOCK(cs_main);
 
             const CChainParams &chainparams = config.GetChainParams();
@@ -1090,15 +1121,15 @@ static RPCHelpMan getblocktemplate() {
                 index_in_template++;
             }
 
+            const bool simplifyGbt = argsman.GetBoolArg("-simplegbt", false);
+
+            UniValue result(UniValue::VOBJ);
             UniValue aux(UniValue::VOBJ);
+            UniValue coinbasetxn(UniValue::VOBJ);
 
-            UniValue minerFundList(UniValue::VARR);
-            for (const auto &fundDestination :
-                 GetMinerFundWhitelist(consensusParams)) {
-                minerFundList.push_back(
-                    EncodeDestination(fundDestination, config));
-            }
-
+            // Compute the miner fund parameters
+            const auto minerFundWhitelist =
+                GetMinerFundWhitelist(consensusParams);
             int64_t minerFundMinValue = 0;
             if (IsAxionEnabled(consensusParams, pindexPrev)) {
                 minerFundMinValue =
@@ -1107,30 +1138,61 @@ static RPCHelpMan getblocktemplate() {
                             SATOSHI);
             }
 
-            UniValue minerFund(UniValue::VOBJ);
-            minerFund.pushKV("addresses", minerFundList);
-            minerFund.pushKV("minimumvalue", minerFundMinValue);
-
-            UniValue coinbasetxn(UniValue::VOBJ);
-            coinbasetxn.pushKV("minerfund", minerFund);
-
+            // Compute the staking reward parameters
             std::vector<CScript> stakingRewardsPayoutScripts;
+            int64_t stakingRewardsAmount =
+                GetStakingRewardsAmount(coinbasevalue) / SATOSHI;
             if (node.avalanche &&
-                IsStakingRewardsActivated(consensusParams, pindexPrev) &&
-                node.avalanche->getStakingRewardWinners(
-                    pindexPrev->GetBlockHash(), stakingRewardsPayoutScripts)) {
-                UniValue stakingRewards(UniValue::VOBJ);
-                UniValue stakingRewardsPayoutScriptObj(UniValue::VOBJ);
-                ScriptPubKeyToUniv(stakingRewardsPayoutScripts[0],
-                                   stakingRewardsPayoutScriptObj,
-                                   /*fIncludeHex=*/true);
-                stakingRewards.pushKV("payoutscript",
-                                      stakingRewardsPayoutScriptObj);
-                stakingRewards.pushKV(
-                    "minimumvalue",
-                    int64_t(GetStakingRewardsAmount(coinbasevalue) / SATOSHI));
+                IsStakingRewardsActivated(consensusParams, pindexPrev)) {
+                if (!node.avalanche->getStakingRewardWinners(
+                        pindexPrev->GetBlockHash(),
+                        stakingRewardsPayoutScripts)) {
+                    stakingRewardsPayoutScripts.clear();
+                }
+            }
 
-                coinbasetxn.pushKV("stakingrewards", stakingRewards);
+            if (simplifyGbt) {
+                UniValue minerFund(UniValue::VOBJ);
+                if (!minerFundWhitelist.empty()) {
+                    minerFund.pushKV("script",
+                                     HexStr(GetScriptForDestination(
+                                         *minerFundWhitelist.begin())));
+                    minerFund.pushKV("amount", minerFundMinValue);
+                }
+                result.pushKV("minerfund", minerFund);
+
+                if (!stakingRewardsPayoutScripts.empty()) {
+                    UniValue stakingRewards(UniValue::VOBJ);
+                    stakingRewards.pushKV(
+                        "script", HexStr(stakingRewardsPayoutScripts[0]));
+                    stakingRewards.pushKV("amount", stakingRewardsAmount);
+                    result.pushKV("stakingrewards", stakingRewards);
+                }
+            } else {
+                UniValue minerFund(UniValue::VOBJ);
+                UniValue minerFundList(UniValue::VARR);
+                for (const auto &fundDestination : minerFundWhitelist) {
+                    minerFundList.push_back(
+                        EncodeDestination(fundDestination, config));
+                }
+
+                minerFund.pushKV("addresses", minerFundList);
+                minerFund.pushKV("minimumvalue", minerFundMinValue);
+
+                coinbasetxn.pushKV("minerfund", minerFund);
+
+                if (!stakingRewardsPayoutScripts.empty()) {
+                    UniValue stakingRewards(UniValue::VOBJ);
+                    UniValue stakingRewardsPayoutScriptObj(UniValue::VOBJ);
+                    ScriptPubKeyToUniv(stakingRewardsPayoutScripts[0],
+                                       stakingRewardsPayoutScriptObj,
+                                       /*fIncludeHex=*/true);
+                    stakingRewards.pushKV("payoutscript",
+                                          stakingRewardsPayoutScriptObj);
+                    stakingRewards.pushKV("minimumvalue", stakingRewardsAmount);
+
+                    coinbasetxn.pushKV("stakingrewards", stakingRewards);
+                }
             }
 
             arith_uint256 hashTarget =
@@ -1141,7 +1203,6 @@ static RPCHelpMan getblocktemplate() {
             aMutable.push_back("transactions");
             aMutable.push_back("prevblock");
 
-            UniValue result(UniValue::VOBJ);
             result.pushKV("capabilities", aCaps);
 
             result.pushKV("version", pblock->nVersion);