+++ test/functional/abc_mining_basic.py
@@ -5,6 +5,7 @@
 Tests for Bitcoin ABC mining RPCs
 """
 
+from test_framework.address import SCRIPT_UNSPENDABLE_HEX
 from test_framework.cdefs import (
     BLOCK_MAXBYTES_MAXSIGCHECKS_RATIO,
     DEFAULT_MAX_BLOCK_SIZE,
@@ -18,6 +19,7 @@ MINER_FUND_ADDR = "ecregtest:prfhcnyqnl5cgrnmlfmms675w93ld7mvvq9jcw0zsn"
 MINER_FUND_LEGACY_ADDR = "2NCXTUCFd1Q3EteVpVVDTrBBoKqvMPAoeEn"
 
 MINER_FUND_RATIO = 32
+STAKING_REWARD_RATIO = 10
 
 
 class AbcMiningRPCTest(BitcoinTestFramework):
@@ -84,6 +86,55 @@ class AbcMiningRPCTest(BitcoinTestFramework):
             }
         )
 
+    def test_simplified_gbt(self):
+        self.restart_node(
+            0, extra_args=["-enableminerfund", "-avalanchestakingrewards"]
+        )
+
+        node = self.nodes[0]
+        node.add_p2p_connection(P2PInterface())
+
+        # Force a staking reward winner so we have an block template output
+        node.setstakingreward(node.getbestblockhash(), SCRIPT_UNSPENDABLE_HEX)
+
+        gbt = node.getblocktemplate()
+        assert "minerfund" not in gbt
+        assert "stakingrewards" not in gbt
+        assert "minerfund" in gbt["coinbasetxn"]
+        assert "stakingrewards" in gbt["coinbasetxn"]
+
+        self.restart_node(
+            0,
+            extra_args=["-enableminerfund", "-avalanchestakingrewards", "-simplegbt"],
+        )
+
+        node.add_p2p_connection(P2PInterface())
+
+        node.setstakingreward(node.getbestblockhash(), SCRIPT_UNSPENDABLE_HEX)
+
+        gbt = node.getblocktemplate()
+        assert "minerfund" in gbt
+        assert "stakingrewards" in gbt
+        assert "minerfund" not in gbt["coinbasetxn"]
+        assert "stakingrewards" not in gbt["coinbasetxn"]
+
+        coinbase = node.getblock(node.getbestblockhash(), 2)["tx"][0]
+        block_reward = sum([vout["value"] for vout in coinbase["vout"]])
+
+        assert_equal(
+            gbt["minerfund"]["script"],
+            node.validateaddress(MINER_FUND_ADDR)["scriptPubKey"],
+        )
+        assert_equal(
+            gbt["minerfund"]["amount"], block_reward * MINER_FUND_RATIO // 100 * XEC
+        )
+
+        assert_equal(gbt["stakingrewards"]["script"], SCRIPT_UNSPENDABLE_HEX)
+        assert_equal(
+            gbt["stakingrewards"]["amount"],
+            block_reward * STAKING_REWARD_RATIO // 100 * XEC,
+        )
+
     def run_test(self):
         self.run_for_node(
             self.nodes[0],
@@ -93,6 +144,7 @@ class AbcMiningRPCTest(BitcoinTestFramework):
             self.nodes[1],
             MINER_FUND_LEGACY_ADDR,
         )
+        self.test_simplified_gbt()
 
 
 if __name__ == "__main__":